requirements.txt
flake8==7.1.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.1
mccabe==0.7.0
packaging==24.2
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.22.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
watchdog==6.0.0
Werkzeug==3.1.3
